cloud_functions/process_gmail_email/requirements.txt
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
fpdf==1.7.2
